feat(typescript-application-project): create typescript-application-project project type
