+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './rust/RustProject';
 export * from './typescript/TypescriptLibraryProject';
+export * from './typescript/TypescriptApplicationProject';
